@@ -4,12 +4,13 @@
   "description": "A lambda function to enable comment-driven actions via github",
   "main": "index.js",
   "scripts": {
-    "eslint": "eslint .",
+    "eslint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "prettier": "prettier --write .",
     "pretest": "prettier --check . && yarn eslint",
     "test": "jest --ci --coverage",
     "test:watch": "jest --watch --notify",
-    "prepare": "husky install"
+    "prepack": "tsc --declarationDir dist --declaration",
+    "prepare": "rm -rf ./dist && husky install"
   },
   "repository": {
     "type": "git",
@@ -25,10 +26,15 @@
   "license": "MIT",
   "private": true,
   "devDependencies": {
+    "@octokit/types": "^6.14.2",
+    "@octokit/webhooks-schemas": "^3.71.2",
+    "@octokit/webhooks-types": "^3.71.2",
     "@types/aws-lambda": "^8.10.76",
     "@types/jest": "^26.0.22",
+    "@types/node": "12.x",
+    "@typescript-eslint/eslint-plugin": "^4.22.1",
+    "@typescript-eslint/parser": "^4.22.1",
     "eslint": "^7.24.0",
-    "eslint-config-airbnb-base": "14.2.1",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jest": "^24.3.5",
@@ -41,17 +47,13 @@
     "typescript": "^4.2.4"
   },
   "engines": {
-    "node": "10.x || 12.x || 14.x"
+    "node": "12.x || 14.x"
   },
   "lint-staged": {
-    "*.js": "eslint --cache --fix",
-    "*.{js,css,md}": "prettier --write"
+    "*.{js,ts}": "eslint --cache --fix",
+    "*.{js,css,md,ts}": "prettier --write"
   },
   "dependencies": {
-    "@octokit/types": "^6.14.2",
-    "@octokit/webhooks-schemas": "^3.71.2",
-    "@octokit/webhooks-types": "^3.71.2",
-    "@types/node": "14.x",
     "aws-lambda": "^1.0.6"
   }
 }
