@@ -1,27 +1,41 @@
-import type { 
-  APIGatewayProxyEvent, 
-  APIGatewayProxyResult 
-} from "aws-lambda";
+import type {
+  APIGatewayProxyEvent,
+  APIGatewayProxyResult,
+  APIGatewayProxyHandler,
+  Context,
+  Callback,
+} from 'aws-lambda';
 import * as https from 'https';
 import * as url from 'url';
-import type { IncomingHttpHeaders } from 'http'
-import type { WebhookEvent, WebhookEventMap, IssueCommentEvent, WebhookEventName, User, PullRequest, IssueComment, Schema, ContentReferenceEvent } from "@octokit/webhooks-types";
-import { ok } from "assert";
+import type { IncomingHttpHeaders } from 'http';
+import type {
+  WebhookEvent,
+  WebhookEventMap,
+  IssueCommentEvent,
+  WebhookEventName,
+  User,
+  PullRequest,
+  IssueComment,
+  Schema,
+} from '@octokit/webhooks-types';
+import { ok } from 'assert';
 import type { Endpoints } from '@octokit/types';
 
 type Unarray<T> = T extends Array<infer U> ? U : T;
 
-type Contents = Unarray<Endpoints["GET /repos/{owner}/{repo}/contents/{path}"]["response"]['data']>;
+type Contents = Unarray<
+  Endpoints['GET /repos/{owner}/{repo}/contents/{path}']['response']['data']
+>;
 
 type Dict<T> = Record<string, T>;
 
 const debug = !!process.env.ENABLE_DEBUG;
 const log = debug
   ? /* istanbul ignore next */ console.log.bind(console) // eslint-disable-line no-console
-  : function log() {};
+  : function log() {}; // eslint-disable-line @typescript-eslint/no-empty-function
 const info = debug
   ? /* istanbul ignore next */ console.info.bind(console) // eslint-disable-line no-console
-  : function info() {};
+  : function info() {}; // eslint-disable-line @typescript-eslint/no-empty-function
 
 log('Loading function');
 
@@ -29,6 +43,14 @@ function assertNotEmpty<T>(thing: T | undefined): T {
   ok(thing);
   return thing;
 }
+
+export type JSONResponse =
+  | { error: string }
+  | ({ success: boolean; triggered: boolean; commented?: boolean } & (
+      | { commentUrl?: string }
+      | { updatedCommentUrl: string }
+      | { message: string }
+    ));
 
 // needs: read_builds, write_builds, read_pipelines
 const BUILDKITE_TOKEN = assertEnv('BUILDKITE_TOKEN');
@@ -52,16 +74,15 @@ class Http404Error extends HttpError {
   }
 }
 
-export const handler = async (
+export const handler: APIGatewayProxyHandler = async (
   event: APIGatewayProxyEvent,
-  context: any /* For legacy testing only */,
-  callback?: Function /* For legacy testing only */,
+  context: Context /* For legacy testing only */,
+  callback?: Callback /* For legacy testing only */,
 ): Promise<APIGatewayProxyResult> => {
-
   log('Received event:', JSON.stringify(event, null, 2));
 
-  const done = (err: Error | null, res?: object) => {
-    const ret = {
+  const done = (err: Error | null, res?: JSONResponse) => {
+    const ret: APIGatewayProxyResult = {
       statusCode: err ? 400 : 200,
       body: err ? JSON.stringify({ error: err.message }) : JSON.stringify(res),
       headers: {
@@ -77,23 +98,25 @@ export const handler = async (
   }
 
   ok(event.headers['X-GitHub-Event']);
-  const currentEventType = event.headers['X-GitHub-Event'] as unknown as WebhookEventName;
+  const currentEventType = (event.headers[
+    'X-GitHub-Event'
+  ] as unknown) as WebhookEventName;
 
   function parseBody<T extends Schema>(event: APIGatewayProxyEvent) {
     ok(event.body);
     log('event body', event.body);
     try {
-      return JSON.parse(event.body) as unknown as T;    
+      return (JSON.parse(event.body) as unknown) as T;
     } catch (e) {
       throw new Error(`Could not parse event body: ${e.message}`);
-    }    
+    }
   }
-
-  
 
   switch (currentEventType) {
     case 'pull_request': {
-      const eventBody = parseBody<WebhookEventMap[typeof currentEventType]>(event);
+      const eventBody = parseBody<WebhookEventMap[typeof currentEventType]>(
+        event,
+      );
       if (eventBody.action !== 'opened') {
         info('PR was not opened, nothing to do here');
         return done(null, { success: true, triggered: false });
@@ -202,8 +225,13 @@ _Note: you can pass [custom environment variables](https://github.com/some-org/s
     }
     case 'issue_comment':
     case 'pull_request_review_comment': {
-      const isIssueComment = (eventName: WebhookEventName,  event: WebhookEvent): event is IssueCommentEvent => eventName === 'issue_comment';
-      const eventBody = parseBody<WebhookEventMap[typeof currentEventType]>(event);
+      const isIssueComment = (
+        eventName: WebhookEventName,
+        event: WebhookEvent,
+      ): event is IssueCommentEvent => eventName === 'issue_comment';
+      const eventBody = parseBody<WebhookEventMap[typeof currentEventType]>(
+        event,
+      );
       const commenter = eventBody.sender.login;
 
       if (commenter === GITHUB_USER) {
@@ -215,7 +243,10 @@ _Note: you can pass [custom environment variables](https://github.com/some-org/s
         info('Comment was deleted, nothing to do here');
         return done(null, { success: true, triggered: false });
       }
-      if (isIssueComment(currentEventType, eventBody) && !eventBody.issue.pull_request) {
+      if (
+        isIssueComment(currentEventType, eventBody) &&
+        !eventBody.issue.pull_request
+      ) {
         info('Request is not coming from a pull request, nothing to do here');
         return done(null, { success: true, triggered: false });
       }
@@ -239,14 +270,19 @@ _Note: you can pass [custom environment variables](https://github.com/some-org/s
       const customEnv = requestedBuildData.env;
       const customEnvKeys = Object.keys(requestedBuildData.env);
       const hasCustomEnv = customEnvKeys.length > 0;
-      requestedBuildData.env = customEnvKeys.reduce<Dict<string>>((ret, key) => {
-        ret[`GH_CONTROL_USER_ENV_${key}`] = requestedBuildData.env[key];
-        return ret;
-      }, {});
+      requestedBuildData.env = customEnvKeys.reduce<Dict<string>>(
+        (ret, key) => {
+          ret[`GH_CONTROL_USER_ENV_${key}`] = requestedBuildData.env[key];
+          return ret;
+        },
+        {},
+      );
 
       return Promise.all([
         isIssueComment(currentEventType, eventBody)
-          ? githubGetPullRequestDetails(assertNotEmpty(eventBody.issue.pull_request?.url))
+          ? githubGetPullRequestDetails(
+              assertNotEmpty(eventBody.issue.pull_request?.url), // TODO: handle empty case gracefully
+            )
           : // pull_request_review_comment events have that information already,
             // no need to query the Github API
             Promise.resolve(eventBody.pull_request as PullRequest),
@@ -259,7 +295,6 @@ _Note: you can pass [custom environment variables](https://github.com/some-org/s
           const senderName = senderData.name;
           const senderEmail = senderData.email ?? undefined;
 
-          // eslint-disable-next-line max-len
           return buildkiteStartBuild(
             requestedBuildData,
             prData,
@@ -330,7 +365,9 @@ ${requestedBuildData.buildNames
         );
     }
     case 'ping': {
-      const eventBody = parseBody<WebhookEventMap[typeof currentEventType]>(event);
+      const eventBody = parseBody<WebhookEventMap[typeof currentEventType]>(
+        event,
+      );
       if (
         !eventBody.hook.active ||
         !eventBody.hook.events ||
@@ -349,7 +386,7 @@ ${requestedBuildData.buildNames
       });
     }
     default:
-      return Promise.reject(new Error(`Unsupported event type "${currentEventType}"`));
+      return done(new Error(`Unsupported event type "${currentEventType}"`));
   }
 };
 
@@ -380,7 +417,9 @@ function urlPart(sshUri: string) {
 /**
  * Retrieves details about a pull request
  */
-async function githubGetPullRequestDetails(prUrl: string /* The fully qualified API URL to a Github comment */) {
+async function githubGetPullRequestDetails(
+  prUrl: string /* The fully qualified API URL to a Github comment */,
+) {
   return githubApiRequest<PullRequest>(prUrl);
 }
 
@@ -416,7 +455,10 @@ async function githubUpdateComment(commentUrl: string, newBody: string) {
  * Variables in the string are given using Bash-like syntax (e.g. $var or
  * ${var}).
  */
-function template(templateString: string, mapping: Dict<string> /* A mapping from a variable name to its value. */): string {
+function template(
+  templateString: string,
+  mapping: Dict<string> /* A mapping from a variable name to its value. */,
+): string {
   function reducer(haystack: string, mappingEntry: [string, string]) {
     const [varName, replacement] = mappingEntry;
     const needle = new RegExp(
@@ -437,7 +479,11 @@ function template(templateString: string, mapping: Dict<string> /* A mapping fro
  * @param {Array} pipeline An array of pipeline objects.
  * @return {Promise} A promise resolving to an array of strings.
  */
-function fetchDocumentationLinkMds(prData: PullRequest, orgSlug: string, pipelines: Pipeline[]) {
+function fetchDocumentationLinkMds(
+  prData: PullRequest,
+  orgSlug: string,
+  pipelines: Pipeline[],
+) {
   return Promise.all(
     pipelines.map((pipeline) =>
       fetchDocumentationLinkMd(prData, BUILDKITE_ORG_NAME, pipeline),
@@ -453,7 +499,11 @@ function fetchDocumentationLinkMds(prData: PullRequest, orgSlug: string, pipelin
  * @param {object} pipeline An array of pipeline objects.
  * @return {Promise} A promise resolving to a string.
  */
-async function fetchDocumentationLinkMd(prData: PullRequest, orgSlug: string, pipeline: Pipeline) {
+async function fetchDocumentationLinkMd(
+  prData: PullRequest,
+  orgSlug: string,
+  pipeline: Pipeline,
+) {
   const documentationUrl = await fetchDocumentationUrl(
     prData,
     orgSlug,
@@ -478,7 +528,11 @@ async function fetchDocumentationLinkMd(prData: PullRequest, orgSlug: string, pi
  * @param {object} pipeline An array of pipeline objects.
  * @return {Promise} A promise resolving to a URL string or null.
  */
-async function fetchDocumentationUrl(prData: PullRequest, orgSlug: string, pipeline: Pipeline): Promise<string | null> {
+async function fetchDocumentationUrl(
+  prData: PullRequest,
+  orgSlug: string,
+  pipeline: Pipeline,
+): Promise<string | null> {
   const docOverrideUrl = (pipeline.env || {}).GH_CONTROL_README_URL;
   if (docOverrideUrl) {
     const mapping = {
@@ -529,7 +583,11 @@ async function fetchDocumentationUrl(prData: PullRequest, orgSlug: string, pipel
  * @param {object} pipeline An array of pipeline objects.
  * @return {string} A link.
  */
-function getDocumentationCreationLink(prData: PullRequest, orgSlug: string, pipeline: Pipeline) {
+function getDocumentationCreationLink(
+  prData: PullRequest,
+  orgSlug: string,
+  pipeline: Pipeline,
+) {
   return (
     '' +
     `https://github.com/${prData.head.repo.full_name}` +
@@ -560,7 +618,11 @@ function zip<S, T>(xs: S[], ys: T[]): [S, T][] {
  * @param {?object} requestBody The JSON body to send to Github
  * @return {Promise} A promise resolving to the decoded JSON data of the Github API response
  */
-async function githubApiRequest<T>(ghUrl: string, additionalOptions?: Partial<https.RequestOptions>, requestBody?: object): Promise<T> {
+async function githubApiRequest<T>(
+  ghUrl: string,
+  additionalOptions?: Partial<https.RequestOptions>,
+  requestBody?: Record<string, unknown>,
+): Promise<T> {
   const options = Object.assign(
     url.parse(ghUrl),
     {
@@ -599,12 +661,16 @@ async function buildkiteReadPipelines() {
   return pipelines;
 }
 
+/**
+ * Incomplete (e.g. only the subset needed by this lambda) type descriptor for the response
+ * of https://buildkite.com/docs/apis/rest-api/pipelines
+ */
 type Pipeline = {
-  slug: string,
-  repository: string,
-  env?: Dict<string>,
-  description?: string,
-}
+  slug: string;
+  repository: string;
+  env?: Dict<string>;
+  description?: string;
+};
 
 /**
  * Starts a buildkite build
@@ -622,7 +688,7 @@ type Pipeline = {
  * @return {Promise} A promise resolving to the decoded JSON data of the Buildkite API response
  */
 async function buildkiteStartBuild(
-  buildData: { buildNames: string[], env: NodeJS.ProcessEnv },
+  buildData: { buildNames: string[]; env: NodeJS.ProcessEnv },
   prData: PullRequest,
   requester: string,
   commentUrl: string,
@@ -665,11 +731,13 @@ async function buildkiteStartBuild(
   );
 }
 
-// see https://buildkite.com/docs/apis/rest-api/builds#list-all-builds
+/**
+ * Incomplete type descriptor of the https://buildkite.com/docs/apis/rest-api/builds#list-all-builds endpoint responses
+ */
 type Build = {
-  web_url: string,
-  number: number,
-  scheduled_at: string,
+  web_url: string;
+  number: number;
+  scheduled_at: string;
 };
 
 /**
@@ -684,7 +752,11 @@ type Build = {
  *                   where body is the decoded JSON data of the Buildkite API response and headers
  *                   is a map
  */
-async function buildkiteApiRequest<T>(apiPathNoOrg: string, additionalOptions? : Partial<https.RequestOptions>, body? : object) {
+async function buildkiteApiRequest<T>(
+  apiPathNoOrg: string,
+  additionalOptions?: Partial<https.RequestOptions>,
+  body?: Record<string, unknown>,
+) {
   const options = {
     hostname: 'api.buildkite.com',
     path: `/v2/organizations/${BUILDKITE_ORG_NAME}/${apiPathNoOrg}`,
@@ -707,7 +779,10 @@ async function buildkiteApiRequest<T>(apiPathNoOrg: string, additionalOptions? :
  *                   where body is the decoded JSON data of the endpoint response and headers
  *                   is a map
  */
-async function jsonRequest<T = object>(options: Partial<https.RequestOptions>, jsonBody?: object): Promise<{ body: T, headers: IncomingHttpHeaders }> {
+async function jsonRequest<T = Record<string, unknown>>(
+  options: Partial<https.RequestOptions>,
+  jsonBody?: Record<string, unknown>,
+): Promise<{ body: T; headers: IncomingHttpHeaders }> {
   const localOptions: https.RequestOptions = { ...options };
   localOptions.headers = Object.assign(options.headers || {}, {
     'User-Agent': 'githubHook/2.0.1', // Note: this is mandatory for the GitHub API
@@ -809,7 +884,7 @@ const buildTriggerRegex = new RegExp(
  * @param {string} commentBody
  * @return {boolean}
  */
-function isTriggerComment(commentBody: string) {
+export function isTriggerComment(commentBody: string): boolean {
   return buildTriggerRegex.test(commentBody);
 }
 
@@ -855,7 +930,9 @@ function parseEnvBlock(envBlock: string) {
  * @param {string} commentBody
  * @return {{buildNames: Array<string>, env: !Object<string,string>}}
  */
-function parseTriggerComment(commentBody: string) {
+export function parseTriggerComment(
+  commentBody: string,
+): { buildNames: string[]; env: Dict<string> } {
   const match = commentBody.match(buildTriggerRegex);
   const pipelinesBlock = match ? match[0] : '';
   const envBlock = match ? match[3] : null;
@@ -866,7 +943,7 @@ function parseTriggerComment(commentBody: string) {
     const pipelineNames = rocket.match(rocketPattern)?.[1];
     if (pipelineNames) {
       const buildName = pipelineNames.match(pipelineNameRegex);
-      if(buildName?.length) {
+      if (buildName?.length) {
         return [...acc, ...buildName];
       }
     }
@@ -878,6 +955,3 @@ function parseTriggerComment(commentBody: string) {
     env: envBlock ? parseEnvBlock(envBlock) : {},
   };
 }
-
-module.exports.isTriggerComment = isTriggerComment;
-module.exports.parseTriggerComment = parseTriggerComment;
